api change and transformation of coordinates
