+++ package.json
@@ -33,11 +33,13 @@
     "express-async-handler": "^1.2.0",
     "express-session": "^1.17.3",
     "parcel": "^2.8.3",
+    "proj4": "^2.9.0",
     "qs": "^6.11.1",
     "ts-node": "^10.9.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.17",
+    "@types/proj4": "^2.5.2",
     "@typescript-eslint/eslint-plugin": "^5.59.0",
     "eslint": "^8.38.0",
     "eslint-config-standard-with-typescript": "^34.0.1",
