--- src/services/api-insee-iris.ts
@@ -1,19 +1,0 @@
-import axios, { type Method } from 'axios'
-
-export async function getIrisCode(lat: string, lon: string) {
-  const url = 'https://pyris.datajazz.io/api/coords?geojson=false&'
-  const config = {
-    method: 'get' as Method,
-    url,
-    params: {
-      lat: lat,
-      lon: lon
-    }
-  }
-  try {
-    const response = await axios(config)
-    return response.data.complete_code
-  } catch (error) {
-    console.error('Error fetching data:', error)
-  }
-}
